+++ TentakelsAttacking2/Saves/config.tac
@@ -18,7 +18,7 @@
         "supernova": true
     },
     "global": {
-        "current_language_name": "english",
+        "current_language_name": "german",
         "game_rounds_current": 50,
         "game_rounds_max": 100,
         "game_rounds_min": 20
@@ -39,7 +39,7 @@
         "player_shuffle": true
     },
     "sound": {
-        "volume_master": 50.0,
+        "volume_master": 56.0,
         "volume_mute_bool": false
     },
     "version": {
