+++ TentakelsAttacking2/TentakelsAttacking2.vcxproj
@@ -234,10 +234,12 @@
     <ClInclude Include="UI\Elements\ColorPicker\public\ColorPickerCell.h" />
     <ClInclude Include="Helper\public\HGeneral.h" />
     <ClInclude Include="UI\public\HoverRender.h" />
+    <ClInclude Include="UI\Scene\MainScenes\public\AppSettings.h" />
     <ClInclude Include="UI\Scene\MainScenes\public\Credits.h" />
+    <ClInclude Include="UI\Scene\MainScenes\public\GameSettings.h" />
     <ClInclude Include="UI\Scene\MainScenes\public\MainScene.h" />
     <ClInclude Include="UI\Scene\MainScenes\public\NewGameParameter.h" />
-    <ClInclude Include="UI\Scene\MainScenes\public\Settings.h" />
+    <ClInclude Include="UI\Scene\SupportScenes\public\Settings.h" />
     <ClInclude Include="UI\Scene\MainScenes\public\UpdateEvaluation.h" />
     <ClInclude Include="UI\Scene\MainScenes\public\ValidateGalaxy.h" />
     <ClInclude Include="UI\public\PopUpManager.h" />
@@ -336,7 +338,8 @@
     <ClCompile Include="UI\Elements\ColorPicker\private\ColorPickerCell.cpp" />
     <ClCompile Include="UI\private\HoverRender.cpp" />
     <ClCompile Include="UI\Scene\MainScenes\private\NewGameParameter.cpp" />
-    <ClCompile Include="UI\Scene\MainScenes\private\Settings.cpp" />
+    <ClCompile Include="UI\Scene\MainScenes\private\AppSettings.cpp" />
+    <ClCompile Include="UI\Scene\SupportScenes\private\Settings.cpp" />
     <ClCompile Include="UI\Scene\MainScenes\private\Credits.cpp" />
     <ClCompile Include="UI\Scene\MainScenes\private\ValidateGalaxy.cpp" />
     <ClCompile Include="UI\private\PopUpManager.cpp" />
