--- TentakelsAttacking2/UI/Scene/MainScenes/public/Settings.h
@@ -1,61 +1,0 @@
-//
-// Purpur Tentakel
-// 01.11.2022
-//
-
-#pragma once
-#include "Scene.h"
-#include "EventListener.hpp"
-
-class SliderAndInputLine;
-
-/**
- * provides a scenes, where the global settings can be set.
- */
-class SettingsScene final : public Scene, public EventListener {
-private:
-	using DropDownButton_ty = std::pair<ClassicButton_ty, bool>;
-	std::shared_ptr<SliderAndInputLine> m_volume; ///< contains the volume slider
-	std::vector<std::pair<Resolution, std::string>> m_rawResolutionEntries; ///< contains die raw resolution information
-	std::pair<DropDownButton_ty, DropDownButton_ty> m_resolutionDropDownBtn; ///< contains the two buttons that are hidden by the resolution drop down
-	DropDown_ty m_resolutionDropDown{ nullptr }; ///< contains the resoltuion drop down
-	std::pair<DropDownButton_ty, DropDownButton_ty> m_languageDropDownBtn; ///< contains the two buttons that are hidden by the language drop down
-	DropDown_ty m_languageDropDown{ nullptr }; ///< contains the language drop down
-	/**
-	 * initializes all ui elements.
-	 * connects the actions.
-	 */
-	void Initialize();
-
-	/**
-	 * returns the strings out of the raw resolution data.
-	 */
-	std::vector<std::string> GetStringsFromResolutionEntries() const;
-	/**
-	 * returns index from resolution enum.
-	 */
-	int GetIndexFromResolution(Resolution resolution) const;
-
-public:
-	/**
-	 * ctor.
-	 * only initialization.
-	 */
-	SettingsScene(Vector2 resolution);
-	~SettingsScene();
-
-	/**
-	 * updates all elements in elements.
-	 */
-	void CheckAndUpdate(Vector2 const& mousePosition, AppContext_ty_c appContext) override;
-	/**
-	 * renders all elements in elements and elements out update.
-	 */
-	void Render(AppContext_ty_c appContext) override;
-	/**
-	 * resizes all elements in elements and elements out update.
-	 */
-	void Resize(Vector2 resolution, AppContext_ty_c appContext) override;
-
-	void OnEvent(Event const& event) override;
-};
